stepconf -fix a spindle-at-speed regression.
The signal spindle-at-speed was left unconnected
in certain combinations of spindle feedback
with no pyvcp display.
This makes it always set True unless:
spindle encoder feedback and
pyvcp display and
use spinndle-at-speed
then it use the near component.
It should use the near component ehn ther is feedback
even if ther is no pyvcp display but we will fix that later.